deleted and renamed files (MVP)
